feat(dashboard): move approver profile modal to sidebar
